ML Projects/DS_AI_Bot_Interviewer/requirements.txt
langchain
streamlit
Pillow
python-dotenv
ipykernel
transformers
ipywidgets
auto-gptq
optimum
nvidia-pyindex
-e .